Rename ui/widget to ui/composable, move sync widget into ui/widget; adapt strings
