+++ app/src/main/kotlin/at/bitfire/davdroid/ui/AppSettingsActivity.kt
@@ -68,11 +68,11 @@ import at.bitfire.davdroid.settings.Settings
 import at.bitfire.davdroid.settings.SettingsManager
 import at.bitfire.davdroid.ui.intro.BatteryOptimizationsPage
 import at.bitfire.davdroid.ui.intro.OpenSourcePage
-import at.bitfire.davdroid.ui.widget.EditTextInputDialog
-import at.bitfire.davdroid.ui.widget.MultipleChoiceInputDialog
-import at.bitfire.davdroid.ui.widget.Setting
-import at.bitfire.davdroid.ui.widget.SettingsHeader
-import at.bitfire.davdroid.ui.widget.SwitchSetting
+import at.bitfire.davdroid.ui.composable.EditTextInputDialog
+import at.bitfire.davdroid.ui.composable.MultipleChoiceInputDialog
+import at.bitfire.davdroid.ui.composable.Setting
+import at.bitfire.davdroid.ui.composable.SettingsHeader
+import at.bitfire.davdroid.ui.composable.SwitchSetting
 import at.bitfire.davdroid.util.PermissionUtils
 import at.bitfire.ical4android.TaskProvider
 import dagger.hilt.android.AndroidEntryPoint
