+++ app/src/main/kotlin/at/bitfire/davdroid/ui/DebugInfoActivity.kt
@@ -83,7 +83,7 @@ import at.bitfire.davdroid.resource.LocalAddressBook
 import at.bitfire.davdroid.settings.AccountSettings
 import at.bitfire.davdroid.settings.SettingsManager
 import at.bitfire.davdroid.syncadapter.BaseSyncWorker
-import at.bitfire.davdroid.ui.widget.CardWithImage
+import at.bitfire.davdroid.ui.composable.CardWithImage
 import at.bitfire.ical4android.TaskProvider
 import at.bitfire.ical4android.TaskProvider.ProviderName
 import at.techbee.jtx.JtxContract
