+++ app/src/main/kotlin/at/bitfire/davdroid/ui/Theme.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalUriHandler
-import at.bitfire.davdroid.ui.widget.SafeAndroidUriHandler
+import at.bitfire.davdroid.ui.composable.SafeAndroidUriHandler
 import com.google.accompanist.themeadapter.material.MdcTheme
 
 val primaryGreen = Color(0xff7cb342)
