+++ app/src/main/kotlin/at/bitfire/davdroid/ui/account/AccountSettingsActivity.kt
@@ -69,12 +69,12 @@ import at.bitfire.davdroid.settings.SettingsManager
 import at.bitfire.davdroid.syncadapter.OneTimeSyncWorker
 import at.bitfire.davdroid.syncadapter.Syncer
 import at.bitfire.davdroid.ui.AppTheme
-import at.bitfire.davdroid.ui.widget.ActionCard
-import at.bitfire.davdroid.ui.widget.EditTextInputDialog
-import at.bitfire.davdroid.ui.widget.MultipleChoiceInputDialog
-import at.bitfire.davdroid.ui.widget.Setting
-import at.bitfire.davdroid.ui.widget.SettingsHeader
-import at.bitfire.davdroid.ui.widget.SwitchSetting
+import at.bitfire.davdroid.ui.composable.ActionCard
+import at.bitfire.davdroid.ui.composable.EditTextInputDialog
+import at.bitfire.davdroid.ui.composable.MultipleChoiceInputDialog
+import at.bitfire.davdroid.ui.composable.Setting
+import at.bitfire.davdroid.ui.composable.SettingsHeader
+import at.bitfire.davdroid.ui.composable.SwitchSetting
 import at.bitfire.davdroid.util.PermissionUtils
 import at.bitfire.ical4android.TaskProvider
 import at.bitfire.vcard4android.GroupMethod
