+++ app/src/main/kotlin/at/bitfire/davdroid/ui/composable/InputDialogs.kt
@@ -1,4 +1,4 @@
-package at.bitfire.davdroid.ui.widget
+package at.bitfire.davdroid.ui.composable
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
