+++ app/src/main/kotlin/at/bitfire/davdroid/ui/composable/PermissionSwitchRow.kt
@@ -1,4 +1,4 @@
-package at.bitfire.davdroid.ui.widget
+package at.bitfire.davdroid.ui.composable
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
