+++ app/src/main/kotlin/at/bitfire/davdroid/ui/widget/SyncButtonWidgetReceiver.kt
@@ -1,4 +1,4 @@
-package at.bitfire.davdroid.widget
+package at.bitfire.davdroid.ui.widget
 
 import androidx.glance.appwidget.GlanceAppWidget
 import androidx.glance.appwidget.GlanceAppWidgetReceiver
